Rename `stack_return` to `compiled_ducky`
