+++ ducklingscript/compiler/__init__.py
@@ -28,5 +28,5 @@ from .stack import Stack
 from .tokenization import Tokenizer, token_return_types
 from .sourcemapping import SourceMap
 from .environments.variable_environment import Null
-from .stack_return import StdOutData
+from .compiled_ducky import StdOutData
 from .compile_options import CompileOptions
