+++ ducklingscript/compiler/commands/bases/block_command.py
@@ -3,7 +3,7 @@ from .doc_command import ArgReqType, ComDoc
 from ducklingscript.compiler.environments.environment import Environment
 from ...errors import InvalidArgumentsError
 from ducklingscript.compiler.pre_line import PreLine
-from ducklingscript.compiler.stack_return import CompiledDucky
+from ducklingscript.compiler.compiled_ducky import CompiledDucky
 from ...tokenization import Tokenizer, token_return_types
 from .base_command import BaseCommand
 from abc import abstractmethod
