+++ ducklingscript/compiler/commands/exist.py
@@ -1,7 +1,7 @@
 from .bases.doc_command import ArgReqType
 from ..errors import GeneralError
 from ducklingscript.compiler.pre_line import PreLine
-from ducklingscript.compiler.stack_return import CompiledDucky
+from ducklingscript.compiler.compiled_ducky import CompiledDucky
 from .bases import ArgLine, SimpleCommand
 
 desc = """
