+++ ducklingscript/compiler/commands/func.py
@@ -1,5 +1,5 @@
 from ducklingscript.compiler.pre_line import PreLine
-from ducklingscript.compiler.stack_return import CompiledDucky
+from ducklingscript.compiler.compiled_ducky import CompiledDucky
 from .bases import BlockCommand, Example
 
 desc = """
