+++ ducklingscript/compiler/commands/if_command.py
@@ -2,7 +2,7 @@ from .bases.doc_command import ArgReqType
 from ducklingscript.compiler.environments.variable_environment import Null
 from ..errors import InvalidArgumentsError
 from ducklingscript.compiler.pre_line import PreLine
-from ducklingscript.compiler.stack_return import CompiledDucky
+from ducklingscript.compiler.compiled_ducky import CompiledDucky
 from .bases import BlockCommand, Example
