+++ ducklingscript/compiler/commands/pass_command.py
@@ -1,4 +1,4 @@
-from ducklingscript.compiler.stack_return import CompiledDucky
+from ducklingscript.compiler.compiled_ducky import CompiledDucky
 from ducklingscript.compiler.pre_line import PreLine
 from .bases import ArgLine, SimpleCommand, ArgReqType
