+++ ducklingscript/compiler/commands/while_loop.py
@@ -1,7 +1,7 @@
 from typing import Any
 
 from ducklingscript.compiler.pre_line import PreLine
-from ducklingscript.compiler.stack_return import StackReturnType, CompiledDucky
+from ducklingscript.compiler.compiled_ducky import StackReturnType, CompiledDucky
 from .bases import BlockCommand, Example
 from ..tokenization import Tokenizer
 from ..errors import ExceededLimitError
